backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
python-decouple
python-dotenv
django-cors-headers
django-storages
boto3
whitenoise
gunicorn
django-filter
razorpay
Pillow
pydub
google-auth
